+++ drizzle/0003_hesitant_dormammu.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "easyinvoice_request" ADD COLUMN "originalRequestPaymentReference" text;
